Removed WindowsAuthentication listener.
Reason: This should only be enabled on the WebSite because otherwise the
WebSite should retrieve and then pass along a valid Kerberos token.

Which is troublesome... since token can expire and only 1 hop is allowed.
JWT is designed for this scenario - using the right tool for the job.

Cons: Both the WebSite and the WebAPI must have identical secret keys,
so they can decode/encode valid JWT's.