Moved option config values for JWTClaimsRetriever
And also fixed a logical bug when retrieving a user.
Also realized that writing a test will be impossible, due to tight
coupling with PrincipalContext and UserPrincipal classes (static methods).
These methods provide access to AD and will be ackward to refactor out.

We take a leap of faith for now and assume that the code works...