FIXED! weird bug! must not use async too much.
Symptom:
Calling next.Invoke(environment) raises ObjectDisposedException, due to
IOwinContext being disposed too early in the lifecycle of the object.

More specifically it was disposed during the call to the database.

Fix:
Changing the async call, to the database, to a synchronous call
".Result()" fixes the issue. Explanation must be that somehow other
stuff is being called behind the scenes and thus disposes the
OwinContext prematurely.