feat: Use TopShelf library for service management
Windows Service management is cumbersome to maintain manually.
You'd need a separate project with the correct settings and thereafter
you'd need to properly install the service with the correct user etc.

With TopShelf you just need to do the configuration in code and then you
are good to go! You also only need 1 project.