feat: Suppress warning when property and variable naming is the same, which could be writting as single word
Reason for supression is that we want to keep the variable name and the property name separate even though
they currently have the same name. Because if we want to change the variable name in the future we will also change
the property name accidentally (tight coupled).

TL;DR: We want to be able to refactor without complications.